Refactored to be more ELHIVy:
Changed elhiv to be just l for prepending function names. This is
where functions will go that do not require further naming beyond
the function name. It is like an area for global functions. This is
as opposed to the function in the 's class' which is in the l_s.vim
file. The l just means language which refers to this library, and
the s is the name of the class. By prepending all my classes with
'l_', this prevents collisions with other function names.