Refactored:
Much shorter syntax/names for constructing classes. The main thing
hard about achieving this is having names that don't clash with
other plugins or functions. That is why before everything started
with l_. I am no longer doing that, which does open up the
possibility of clashing with other names, but I think it is worth
it. I am at least reducing the chances of that happening by having
this code be sourced from inside the user's config file. This means
that my function names need to start with a capital letter now,
which is the only reason they do now.