Started a lower level language layer inside the autoload directory:
I started having too many problems with having classes inside the
elhiv file and the elhiv file trying to source all the right files
in the right order. In order to resolve this, I am building a lower
lever language layer (library) inside the autoload directory that
will serve as the base layer. In more concrete terms, the
elhiv.vim file should only use classes and functions from inside
the autoload directory to source the rest of the plugin. The idea
is to keep the lowest layer as simple as possible and let a more
complex layer build itself on top of that. Essentially for now the
rest of the library that other plugins wil build themselves upon
is this higher level layer. Other plugins are not recommended to
use the classes and functions inside the autoload directory, but
there is nothing wrong with doing so. If a concept from the
autoload directory is also generally useful for more than just
building this library, a wrapper class should be added to normal
classes in the library. These will start out as wrapper classes,
but may end up being more than that. It may prove useful to make
the concepts and names slightly different between these two layers.
For now the Null class is an example of a good wrapper class. If
Dir ends up in the base language it may be better for it to be a
dir_info class, so the Dir class can be more than just a wrapper
class without having to embed all of that inside the base class.