Refactored: No longer sourcing files using hard coded paths:
Now all files in the src/ directory are automatically sourced. This
means I was able to detete several lines of code. It also means I
no longer need to remember to go into the elhiv.vim file and add a
new line every time I add a new class to the src/ directory.