Started a tags class and reorganizing structure:
Tests are divided into bootstrap, boot, and other things in the
future. bootstrap is for the autoload tests because the autoload
files are like a bootstrap for putting on the more powerful boot,
which is the code in the src directory. The code in the src dir is
like the base language. Anything in a subdirectory of src, is not
the boot, but it is a higher level language built on top of the
boot.